fix native pickling of __main__ objects by resorting to replacing __main__ in binary output of pickle.dumps
